Change Cell struct Value type from uint16_t to uint32_t for consistency and improved range

+++ src/Grid.h
@@ -8,7 +8,7 @@
 #define CELL_SIZE 30
 
 struct Cell {
-    uint16_t Value;
+    uint32_t Value;
     Color Color;
 };
